Update README.md
This commit was part of my Lesson 05: Lab assignment. It required me to address 4 items in a README.md file following a guided document. These items were my name and email address, a brief description of my GitHub account, a brief description of my repository, followed by a brief description of the steps I used to clone the repository.