Lab7 Index with CSS
This assignment is a simple index or home page titled, "My Personal Introduction". The goal was to create a webpage using HTML for the structure and some light CSS to provide some visual appeal. The color palette was created using a unique AI color palette generator called "colormind.io". It was difficult to find one that would maintain readability as the palette creator was geared for more advanced users who would use the colors for decorative elements such as overlapping shapes or buttons.
I then had to write some HTML about myself which you can see for yourself.  Thanks for reading!